feat(blockchain): enhance block processing and reorganization logic, update transaction validation

+++ crates/btc-core/src/transaction/transaction.rs
@@ -10,7 +10,7 @@ pub struct Transaction {
     pub version: u32,
     pub inputs: Vec<TxInput>,
     pub outputs: Vec<TxOutput>,
-    pub lock_time: u32,
+    pub lock_time: u64,
 }
 
 impl BitcoinSerialize for Transaction {
